# specs/002-react-admin-frontend/contracts/openapi-admin.yaml
openapi: 3.0.3
info:
  title: Admin API Endpoints
  description: Backend API endpoints for administrative operations in multi-tenant system. Frontend-agnostic APIs supporting any client framework.
  version: 1.0.0
  contact:
    name: Backend API Team
  license:
    name: MIT
servers:
  - url: /api/v1
    description: API v1 base path (admin endpoints)
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Policy:
      type: object
      required: [policy_id, tenant_id, name, resource, action, effect, roles, priority, is_active, created_at, updated_at, created_by, updated_by]
      properties:
        policy_id:
          type: string
          format: uuid
        tenant_id:
          type: string
          format: uuid
        name:
          type: string
          description: Unique within tenant
        resource:
          type: string
          description: Resource type (users, tenants, policies, etc.)
        action:
          type: string
          description: Action type (create, read, update, delete)
        conditions:
          type: object
          properties:
            resource_owner:
              type: boolean
            same_tenant:
              type: boolean
            custom_rules:
              type: object
        effect:
          type: string
          enum: [ALLOW, DENY]
        roles:
          type: array
          items:
            type: string
        priority:
          type: integer
          minimum: 0
          maximum: 1000
        is_active:
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        created_by:
          type: string
          format: uuid
        updated_by:
          type: string
          format: uuid
    FeatureFlag:
      type: object
      required: [flag_id, name, is_enabled, created_at, updated_at, created_by, updated_by]
      properties:
        flag_id:
          type: string
          format: uuid
        tenant_id:
          type: string
          format: uuid
          nullable: true
          description: null for global flags
        name:
          type: string
          description: Unique within tenant scope
        is_enabled:
          type: boolean
        description:
          type: string
          nullable: true
        rollout_percentage:
          type: integer
          minimum: 0
          maximum: 100
          nullable: true
        target_users:
          type: array
          items:
            type: string
            format: uuid
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        created_by:
          type: string
          format: uuid
        updated_by:
          type: string
          format: uuid
    Invitation:
      type: object
      required: [invitation_id, tenant_id, email, roles, invited_by, status, expires_at, invitation_token, created_at, updated_at]
      properties:
        invitation_id:
          type: string
          format: uuid
        tenant_id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        roles:
          type: array
          items:
            type: string
        invited_by:
          type: string
          format: uuid
        status:
          type: string
          enum: [pending, accepted, expired, revoked]
        expires_at:
          type: string
          format: date-time
        accepted_at:
          type: string
          format: date-time
          nullable: true
        accepted_by:
          type: string
          format: uuid
          nullable: true
        invitation_token:
          type: string
          description: Secure token for invitation acceptance
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    ListResponse:
      type: object
      required: [data, pagination]
      properties:
        data:
          type: array
          items: {}
        pagination:
          type: object
          required: [page, per_page, total, total_pages]
          properties:
            page:
              type: integer
              minimum: 1
            per_page:
              type: integer
              minimum: 1
              maximum: 100
            total:
              type: integer
              minimum: 0
            total_pages:
              type: integer
              minimum: 0
    SingleResponse:
      type: object
      required: [data]
      properties:
        data: {}
...
    CreatePolicyRequest:
      type: object
      required: [name, resource, action, effect, roles]
      properties:
        name:
          type: string
        resource:
          type: string
        action:
          type: string
        conditions:
          type: object
        effect:
          type: string
          enum: [ALLOW, DENY]
        roles:
          type: array
          items:
            type: string
        priority:
          type: integer
          minimum: 0
          maximum: 1000
          default: 100
        is_active:
          type: boolean
          default: true
    CreateFeatureFlagRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
        tenant_id:
          type: string
          format: uuid
          nullable: true
        is_enabled:
          type: boolean
          default: false
        description:
          type: string
        rollout_percentage:
          type: integer
          minimum: 0
          maximum: 100
        target_users:
          type: array
          items:
            type: string
            format: uuid
    CreateInvitationRequest:
      type: object
      required: [email, roles]
      properties:
        email:
          type: string
          format: email
        roles:
          type: array
          items:
            type: string
        expires_in_hours:
          type: integer
          minimum: 1
          maximum: 168
          default: 72
  parameters:
    TenantIdQuery:
      name: tenant_id
      in: query
      description: Filter by tenant ID (required for superadmin cross-tenant operations)
      schema:
        type: string
        format: uuid
    PageQuery:
      name: page
      in: query
      description: Page number
      schema:
        type: integer
        minimum: 1
        default: 1
    PerPageQuery:
      name: per_page
      in: query
      description: Items per page
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 25
...
paths:
...
  /policies:
    get:
      summary: List policies
      description: List policies within authorized tenant scope
      parameters:
        - $ref: '#/components/parameters/TenantIdQuery'
        - $ref: '#/components/parameters/PageQuery'
        - $ref: '#/components/parameters/PerPageQuery'
        - name: resource
          in: query
          description: Filter by resource type
          schema:
            type: string
        - name: effect
          in: query
          description: Filter by policy effect
          schema:
            type: string
            enum: [ALLOW, DENY]
        - name: is_active
          in: query
          description: Filter by active status
          schema:
            type: boolean
      responses:
        '200':
          description: List of policies
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ListResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/Policy'

    post:
      summary: Create policy
      description: Create a new policy within authorized tenant scope
      parameters:
        - $ref: '#/components/parameters/TenantIdQuery'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePolicyRequest'
      responses:
        '201':
          description: Policy created
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SingleResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Policy'

  # Feature Flag Management Endpoints
  /feature-flags:
    get:
      summary: List feature flags
      description: List feature flags within authorized scope (global + tenant-specific)
      parameters:
        - $ref: '#/components/parameters/TenantIdQuery'
        - $ref: '#/components/parameters/PageQuery'
        - $ref: '#/components/parameters/PerPageQuery'
        - name: is_enabled
          in: query
          description: Filter by enabled status
          schema:
            type: boolean
      responses:
        '200':
          description: List of feature flags
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ListResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/FeatureFlag'

    post:
      summary: Create feature flag
      description: Create a new feature flag
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateFeatureFlagRequest'
      responses:
        '201':
          description: Feature flag created
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SingleResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/FeatureFlag'

  # Invitation Management Endpoints
  /invitations:
    get:
      summary: List invitations
      description: List invitations within authorized tenant scope
      parameters:
        - $ref: '#/components/parameters/TenantIdQuery'
        - $ref: '#/components/parameters/PageQuery'
        - $ref: '#/components/parameters/PerPageQuery'
        - name: status
          in: query
          description: Filter by invitation status
          schema:
            type: string
            enum: [pending, accepted, expired, revoked]
      responses:
        '200':
          description: List of invitations
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ListResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/Invitation'

    post:
      summary: Create invitation
      description: Create a new user invitation within authorized tenant scope
      parameters:
        - $ref: '#/components/parameters/TenantIdQuery'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateInvitationRequest'
      responses:
        '201':
          description: Invitation created
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SingleResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Invitation'
...